Major refactoring
Moving many internal functions into core.js that can be required by both 
npm.js and go.js
