+++ lib/commands/export.js
@@ -5,9 +5,10 @@ const forest = require('../forest');
 const async = require('async');
 
 (async () => {
-  forest.connectDB()
-  await forest.connectIPFS();
+  var db = forest.connectDB()
+  await forest.connectIPFS(db);
 
-  var stats = await forest.exportPackages()
+  var stats = await forest.core.exportPackages(db)
   console.log(stats)
+  await db.close()
 })();
