+++ lib/commands/id.js
@@ -4,7 +4,8 @@
 const forest = require('../forest');
 
 (async () => {
-  forest.connectDB()
-  var ipfsID = await forest.connectIPFS();
+  var db = forest.connectDB()
+  var ipfsID = await forest.connectIPFS(db);
   console.log('Peer ID:',ipfsID.id)
+  await db.close()
 })()
