+++ lib/commands/import.js
@@ -7,8 +7,8 @@ const forest = require('../forest');
 const async = require('async');
 
 (async () => {
-  forest.connectDB()
-  await forest.connectIPFS();
+  var db = forest.connectDB()
+  await forest.connectIPFS(db);
 
   var q = async.queue(async function(task, callback) {
     var parts = task.pkg.split(':')
