+++ lib/commands/packages.js
@@ -4,10 +4,10 @@
 const forest = require('../forest');
 
 (async () => {
-  forest.connectDB()
+  var db = forest.connectDB()
   forest.listPackages().then(async function(packages) {
     packages.forEach(pkg => console.log(pkg.manager, pkg.name, pkg.version));
     console.log('Total: ' + packages.length, 'packages')
-    await forest.closeDB()
+    await db.close()
   })
 })()
