+++ lib/commands/peers.js
@@ -4,11 +4,11 @@
 const forest = require('../forest');
 
 (async () => {
-  forest.connectDB()
+  var db = forest.connectDB()
 
-  forest.listPeers().then(async function(peers) {
+  forest.core.listPeers(db).then(async function(peers) {
     peers.forEach(peer => console.log(peer));
     console.log('Total: ' + peers.length, 'peers')
-    await forest.closeDB()
+    await db.close()
   })
 })();
