+++ lib/commands/preload.js
@@ -7,34 +7,25 @@ const path = require("path");
 const async = require('async');
 
 (async () => {
-  forest.connectDB()
-  await forest.connectIPFS();
+  var db = forest.connectDB()
+  await forest.connectIPFS(db);
 
   var packages = []
 
   console.log('Searching for package-lock.json and go.sum files...')
 
-  var q = async.queue(async function(task, callback) {
+  var q = async.queue(async function(task) {
     const {manager, name, pkg} = task
     console.log("Importing", name, pkg.version)
     if(pkg.resolved){
       try{
         // if npm and sha512 available, hashAlg = sha2-512
-        var id = await forest.addUrltoIPFS(manager, name, pkg.version, pkg.resolved)
+        await forest.importPackage(manager, name, pkg.version, pkg.resolved)
       } catch(e) {
         console.error(name, pkg.version, e)
-        callback();
-      }
-      try{
-        await forest.loadMetadata('npm', name)
-        callback();
-      } catch(e) {
-        console.error("loadMetadata error", name, e)
-        callback();
       }
     } else {
       console.log("URL missing for", name, pkg.version)
-      callback();
     }
   }, 20);
