+++ lib/commands/republish.js
@@ -6,8 +6,8 @@ const forest = require('../forest');
 const async = require('async');
 
 (async () => {
-  forest.connectDB()
-  await forest.connectIPFS();
+  var db = forest.connectDB()
+  await forest.connectIPFS(db);
 
   var cids = {}
 
@@ -20,6 +20,7 @@ const async = require('async');
     console.log(manager, name, pkg.version)
 
     var possibleCID = await forest.fetchAndAddtoIPFS(manager, name, pkg.version, pkg.resolved)
+    console.log('possibleCID', possibleCID)
     console.log('Integrity:', pkg.integrity)
     if(possibleCID){
       console.log(`Expected CID to load from IPFS: ${possibleCID}`)
@@ -29,7 +30,7 @@ const async = require('async');
     }
 
     if(pkg.resolved){
-      var res = await forest.addUrltoIPFS(manager, name, pkg.version, pkg.resolved, hashAlg)
+      var res = await forest.core.addUrltoIPFS(db, manager, name, pkg.version, pkg.resolved, hashAlg)
       console.log('Actual CID from adding to IPFS:', res)
       if(res) { cids[`${manager}:${name}:${pkg.version}`] = res }
     } else {
