+++ lib/commands/seed.js
@@ -2,11 +2,10 @@
 // listen for new packages announced on the IPFS DHT and republish them
 
 const forest = require('../forest');
-const uint8ArrayToString = require('uint8arrays/to-string');
 
 (async () => {
-  forest.connectDB()
-  await forest.connectIPFS();
+  var db = forest.connectDB()
+  await forest.connectIPFS(db);
 
   forest.subscribePackageAnnoucements(forest.seed);
 })();
