+++ lib/commands/server.js
@@ -1,13 +1,14 @@
 // forest daemon
 // start the http server and proxy
 
-const server = require('../server');
+const createServer = require('../server');
 const forest = require('../forest');
 
 (async () => {
-  forest.connectDB()
-  var ipfsID = await forest.connectIPFS();
+  var db = forest.connectDB()
+  var ipfsID = await forest.connectIPFS(db);
   if(ipfsID){
+    server = await createServer(db)
     server.listen(8005)
     forest.subscribePackageAnnoucements()
     forest.watchKnown()
