+++ lib/commands/update.js
@@ -5,8 +5,8 @@ const forest = require('../forest');
 const async = require('async');
 
 (async () => {
-  forest.connectDB()
-  await forest.connectIPFS();
+  var db = forest.connectDB()
+  await forest.connectIPFS(db);
 
   var q = async.queue(async function(pkg, callback) {
     console.log('Checking:', pkg.name)
