+++ lib/commands/verify.js
@@ -5,8 +5,8 @@ const forest = require('../forest');
 const async = require('async');
 
 (async () => {
-  forest.connectDB()
-  await forest.connectIPFS();
+  var db = forest.connectDB()
+  await forest.connectIPFS(db);
 
   var q = async.queue(async function(pkg, callback) {
     var res = await forest.validate(pkg.manager, pkg.name, pkg.version)
@@ -14,11 +14,11 @@ const async = require('async');
       console.log('Verifying:', pkg.manager, pkg.name, pkg.version, '✅')
     } else {
       console.log('Verifying:', pkg.manager, pkg.name, pkg.version, '❌')
-      await forest.removePackage(pkg.manager, pkg.name, pkg.version)
+      // await forest.removePackage(pkg.manager, pkg.name, pkg.version)
     }
 
     callback()
-  }, 20)
+  }, 1)
 
   // TODO allow passing in a manager via cli --manager=go
   forest.listPackages().then(async function(packages) {
