+++ lib/commands/watch.js
@@ -4,8 +4,8 @@
 const forest = require('../forest');
 
 (async () => {
-  forest.connectDB()
-  var ipfsID = await forest.connectIPFS();
+  var db = forest.connectDB()
+  var ipfsID = await forest.connectIPFS(db);
   if(ipfsID){
     forest.watchAll();
   }
