+++ lib/forest.js
@@ -2,21 +2,17 @@ const uint8ArrayToString = require('uint8arrays/to-string')
 const IpfsHttpClient = require('ipfs-http-client')
 const { urlSource } = IpfsHttpClient
 const ipfs = IpfsHttpClient()
-const detectContentType = require('ipfs-http-response/src/utils/content-type')
-const toStream = require('it-to-stream')
 const pipe = require('it-pipe')
 const toIterable = require('stream-to-it')
-const fetch = require('node-fetch')
-const fs = require('fs-extra');
-const path = require('path');
+
 const tmp = require('tmp-promise');
 const {zipDigest, zipManifest} = require('./zip_digest');
-const CID = require('cids')
-const multihash = require('multihashes')
+
 const ssri = require('ssri')
 var crypto = require('crypto');
 const StreamZip = require('node-stream-zip');
 
+const core = require('./managers/core')
 const npm = require('./managers/npm')
 const go = require('./managers/go')
 
@@ -25,32 +21,39 @@ const level = require('level-party')
 const paths = envPaths('forest');
 var db
 
-var pjson = require('../package.json');
-
 const packageAnnoucementsTopic = 'forest'
 
 var ipfsID = undefined;
-var ipfsd = undefined;
 
+// db things
 function connectDB() {
   return db = level(paths.data)
 }
 
-function version() {
-  pjson.version
+async function reset() {
+  await db.clear()
 }
 
-async function connectIPFS(){
+function filteredReadStream(start) {
+  return db.createReadStream({gte: start, lt: start+ '~'})
+}
+
+async function closeDB(){
+  return await db.close()
+}
+
+// ipfs things
+async function connectIPFS(db){
   try{
     ipfsID = await ipfs.id()
     console.log('Connected to IPFS')
-    dialPeers()
+    core.dialPeers(db)
     return ipfsID;
   } catch {
     try{
-      ipfsd = await startIPFS()
+      var ipfsd = await core.startIPFS()
       ipfsID = await ipfsd.api.id()
-      dialPeers()
+      core.dialPeers(db)
       return ipfsID;
     } catch(e){
       console.log('ERROR', e)
@@ -60,44 +63,10 @@ async function connectIPFS(){
   }
 }
 
-async function startIPFS() {
-  console.log('Starting IPFS')
 
-  const Ctl = require('ipfsd-ctl');
 
-  const ipfsd = await Ctl.createController({
-      disposable: false,
-      args: '--enable-pubsub-experiment',
-      ipfsHttpModule: require('ipfs-http-client'),
-      ipfsBin: require('go-ipfs').path(),
-      test: false,
-      remote: false
-  })
 
-  fs.removeSync(path.join(ipfsd.path, 'api'))
-
-  var started = await ipfsd.start()
-  console.log('Started IPFS')
-  const id = await ipfsd.api.id()
-  return ipfsd
-}
-
-async function dialPeers() {
-  listPeers().then(async function(peers) {
-    peers.forEach(async function(peer) {
-      try {
-        await ipfs.swarm.connect(`/p2p/${peer}`)
-      } catch (e) {
-        // couldn't find peer, nbd
-      }
-    });
-  })
-}
-
-async function reset() {
-  await db.clear()
-}
-
+// pubsub things
 async function defaultAnnounceCb(msg) {
   try{
     var json = JSON.parse(uint8ArrayToString(msg.data))
@@ -146,6 +115,19 @@ async function defaultAnnounceCb(msg) {
   }
 }
 
+async function subscribePackageAnnoucements(receiveMsg = defaultAnnounceCb) {
+  try {
+    await ipfs.pubsub.subscribe(packageAnnoucementsTopic, receiveMsg)
+    console.log(`Subscribed to '${packageAnnoucementsTopic}' pubsub topic`)
+  } catch(e) {
+    console.error(`Failed to subscribe to '${packageAnnoucementsTopic}' pubsub topic`)
+    console.log("IPFS experimental pubsub feature not enabled. Run daemon with --enable-pubsub-experiment")
+  }
+}
+
+
+
+// commands
 async function seed(msg){
 
   if (ipfsID.id === msg.from) { return } // ignore our own announcements
@@ -159,56 +141,16 @@ async function seed(msg){
 
   // TODO fallback to http if download from IPFS fails or times out
   downloadPackageFromIPFS(json.manager, json.name, json.version, json.cid);
-  // addUrltoIPFS(json.manager, json.name, json.version, json.url)
+  // core.addUrltoIPFS(db, json.manager, json.name, json.version, json.url)
 }
 
-async function subscribePackageAnnoucements(receiveMsg = defaultAnnounceCb) {
-  try {
-    await ipfs.pubsub.subscribe(packageAnnoucementsTopic, receiveMsg)
-    console.log(`Subscribed to '${packageAnnoucementsTopic}' pubsub topic`)
-  } catch(e) {
-    console.error(`Failed to subscribe to '${packageAnnoucementsTopic}' pubsub topic`)
-    console.log("IPFS experimental pubsub feature not enabled. Run daemon with --enable-pubsub-experiment")
-  }
+async function importPackage(manager, name, version, resolved) {
+  await core.addUrltoIPFS(db, manager, name, version,resolved)
+  await core.loadMetadata(db, manager, name)
 }
 
-async function unsubscribePackageAnnoucements() {
-  await ipfs.pubsub.unsubscribe(packageAnnoucementsTopic)
-  console.log(`Unsubscribed from '${packageAnnoucementsTopic}' pubsub topic`)
-}
 
-// has npm specific code
-async function downloadPackageFromRegistry(manager, name, version){
-  var metadata = await loadMetadata(manager, name)
-  if(!metadata) {
-    console.log("Failed to load metadata for", name)
-    return false
-  }
-
-  var versionData = metadata.versions[version]
-
-  if(versionData == null){
-    console.log('Reloading metadata for', name, version)
-    const metadata = await loadMetadataFromRegistry(manager, name)
-    var versionData = metadata.versions[version]
-
-    if(versionData == null){
-      // no known version in registy
-      console.log("Can't find", version, "for", name)
-      return false
-    }
-  }
-
-  var url = versionData.dist.tarball
-  // if sha512 available, hashAlg = sha2-512
-  return await addUrltoIPFS(manager, name, version, url)
-}
-
-async function downloadGoPackageFromRegistry(name, version) {
-  var url = `https://proxy.golang.org/${name}/@v/${version}.zip`
-  return await addUrltoIPFS('go', name, version, url)
-}
-
+// core
 async function fetchAndAddtoIPFS(manager, name, version, url) {
   var hashAlg = 'sha2-256'
   var possibleCID
@@ -218,24 +160,24 @@ async function fetchAndAddtoIPFS(manager, name, version, url) {
   // TODO allow passing in integrity from lockfile
   if(manager === 'npm'){
     // grab the integrity hash from the registry
-    var meta = await loadVersionMetadata(manager, name, version)
+    var meta = await core.loadVersionMetadata(db, manager, name, version)
     var integrity = meta.dist.integrity
     if(integrity){
       var hashAlg = 'sha2-512'
       // guess the cid
-      possibleCID = guessCID(integrity)
+      possibleCID = core.guessCID(integrity)
     }
   }
   if(manager === 'go'){
     var integrity = await go.fetchChecksum(name, version)
-    possibleCID = guessCID(`sha256-${integrity}`)
+    possibleCID = core.guessCID(`sha256-${integrity}`)
   }
 
   if(possibleCID){
     console.log('possibleCID', possibleCID)
 
     // attempt to load it via ipfs
-    var file = await attemptIPFSLoad(possibleCID)
+    var file = await core.attemptIPFSLoad(possibleCID)
 
     if(file){
       // guessed CID was correct and loaded from ipfs
@@ -310,7 +252,7 @@ async function fetchAndAddtoIPFS(manager, name, version, url) {
           var u8 = await zip.generateAsync({type: "uint8array", compression: "DEFLATE",compressionOptions: {level: 9}})
 
 
-          var zipcid = await ipfsAdd(u8)
+          var zipcid = await core.ipfsAdd(u8)
 
 
            cid = zipcid // TODO should be the cid of the zip, not the metafile
@@ -332,7 +274,7 @@ async function fetchAndAddtoIPFS(manager, name, version, url) {
       console.log("Didn't manange to guess and load files via ipfs")
       // if files didn't load
       if(manager === 'go'){
-        var zipCID = await addUrltoIPFS(manager, name, version, url, hashAlg)
+        var zipCID = await core.addUrltoIPFS(db, manager, name, version, url, hashAlg)
 
         var dir = await tmp.dir()
         var zippath = dir.path + '/' + version + '.zip'
@@ -344,25 +286,26 @@ async function fetchAndAddtoIPFS(manager, name, version, url) {
         )
 
         metafile = await zipManifest(zippath)
-        var metafileCID = await ipfsAdd(metafile, hashAlg)
+        var metafileCID = await core.ipfsAdd(metafile, hashAlg)
         var cids = await addZipContentsToIPFS(zippath)
 
         await fs.remove(zippath)
         cid = zipCID
       } else {
         // load via url, add cid to db and announce over pubsub
-        cid = await addUrltoIPFS(manager, name, version, url, hashAlg)
+        cid = await core.addUrltoIPFS(db, manager, name, version, url, hashAlg)
       }
     }
 
   } else {
     // regular url download (only old npms + other possible package managers)
-    cid = await addUrltoIPFS(manager, name, version, url, hashAlg)
+    cid = await core.addUrltoIPFS(db, manager, name, version, url, hashAlg)
   }
 
   return cid
 }
 
+// go specific
 async function addMetalineToIPFS(line){
   try{
 
@@ -373,9 +316,9 @@ async function addMetalineToIPFS(line){
     }
    console.log('Searching for', parts[1])
    var digest = ssri.fromHex(parts[0], 'sha256', {single: true})
-   var maybeFileCID = guessCID(digest)
+   var maybeFileCID = core.guessCID(digest)
 
-   var individualFileContent = await attemptIPFSLoad(maybeFileCID)
+   var individualFileContent = await core.attemptIPFSLoad(maybeFileCID)
 
    if (individualFileContent){
      f.cid = maybeFileCID
@@ -392,6 +335,7 @@ async function addMetalineToIPFS(line){
  }
 }
 
+// go specific
 async function addZipContentsToIPFS(filePath) {
   // TODO needs to be async
   return new Promise((resolve, reject) => {
@@ -402,7 +346,7 @@ async function addZipContentsToIPFS(filePath) {
         if(!entry.isDirectory){
 
           const data = zip.entryDataSync(entry.name);
-          var cid = await ipfsAdd(data, 'sha2-256')
+          var cid = await core.ipfsAdd(data, 'sha2-256')
           cids.push(cid)
           console.log('Adding', entry.name, 'to ipfs', cid.cid.toString())
         }
@@ -423,86 +367,10 @@ async function addZipContentsToIPFS(filePath) {
 });
 }
 
-async function ipfsAdd(data, hashAlg = 'sha2-256') {
-  return await ipfs.add(data, {chunker: 'size-1048576', rawLeaves: true, hashAlg: hashAlg})
-}
-
-async function attemptIPFSLoad(cid){
-  // console.log('cid', cid)
-  // var files = []
-  // try{
-  //   for await (const file of ipfs.get(cid, {timeout: 1*1000})) { // 5 second timeout
-  //     files.push(file)
-  //   }
-  // } catch(e){
-  //   console.error(e)
-  //   // timeout error
-  //   // no peers with content or took too long
-  // }
-
-  try{
-    var url = "http://127.0.0.1:5001/api/v0/cat?arg=/ipfs/"+cid
-    var response = await fetch(url, {method: 'POST'});
-    var text = await response.text();
-    return text
-  }catch(e){
-    console.error(e)
-    return false
-  }
-}
 
 
-function guessCID(integrity) {
-  var sha = ssri.parse(integrity, {single: true})
-  if (sha.algorithm === 'sha512'){
-    var mhash = multihash.fromHexString('1340' + sha.hexDigest())
-    return possibleCID = new CID(1, 'raw', mhash).toString()
-  } else if (sha.algorithm === 'sha256'){
-    var mhash = multihash.fromHexString('1220' + sha.hexDigest())
-    return possibleCID = new CID(1, 'raw', mhash).toString()
-  } else {
-    return false
-  }
-}
 
-async function addUrltoIPFS(manager, name, version, url, hashAlg = 'sha2-256'){
-  try {
-    exists = await db.get(`cid:${manager}:${name}:${version}`)
-  } catch (e) {
-    exists = false
-  }
 
-  if (exists) { return exists }
-
-  try {
-    const file = await ipfsAdd(urlSource(url), hashAlg)
-    console.log('IPFS add:', file.path, file.cid.toString())
-
-    await db.put(`cid:${manager}:${name}:${version}`, file.cid.toString())
-    var size = await setTarballSize(manager, name, version)
-
-    try {
-      ipfs.pubsub.publish('forest', JSON.stringify({
-        url: url,
-        manager: manager,
-        name: name,
-        version: version,
-        path: file.path,
-        cid: file.cid.toString(),
-        forest: pjson.version,
-        size: size
-      }))
-    } catch(e) {
-      console.error('Failed to announce', manager, name, version, 'over pubsub')
-    }
-
-    return file.cid.toString()
-  } catch(e) {
-    console.log('error in ipfs add', url)
-    console.error(e)
-    return false
-  }
-}
 
 function listPackages(manager) {
   if(manager){
@@ -523,10 +391,6 @@ function listPackages(manager) {
       })
   })
   return packages
-}
-
-async function closeDB(){
-  return await db.close()
 }
 
 function listPackageNames(manager) {
@@ -550,19 +414,7 @@ function listPackageNames(manager) {
   })
 }
 
-function listPeers() {
-  return new Promise((resolve, reject) => {
-    var names = []
-    db.createKeyStream({gte: 'repub:', lt: 'repub:~'})
-      .on('data', function (data) {
-        var parts = data.split(':')
-        names.push(parts[4])
-      })
-      .on('end', function () {
-        resolve( [...new Set(names)])
-      })
-  })
-}
+
 
 async function downloadPackageFromIPFS(manager, name, version, cid) {
   try {
@@ -577,315 +429,46 @@ async function downloadPackageFromIPFS(manager, name, version, cid) {
   }
 
   if(manager === 'npm'){
-    return await npmVerify(name, version, cid)
+    return await npm.verify(db, name, version, cid)
   } else if (manager === 'go'){
-    return await goVerify(name, version, cid)
+    return await go.verify(db, name, version, cid)
   } else {
     console.log('Unknown manager:', manager)
   }
 }
 
-async function npmVerify(name, version, cid) {
-  var manager = 'npm'
-  var metadata = await loadMetadata(manager, name)
-  if(!metadata) {
-    console.log("Failed to load metadata for", name)
-    return false
-  }
-  var versionData = metadata.versions[version]
-
-  if (versionData == null) {
-    console.log('Reloading metadata for', name, version)
-    var metadata = await loadMetadataFromRegistry(manager, name)
-    var versionData = metadata.versions[version]
-  }
-
-  if (versionData) {
-    if(versionData.dist.integrity){
-      var integrity = versionData.dist.integrity
-    } else {
-      var integrity = ssri.fromHex(versionData.dist.shasum, 'sha1').toString()
-    }
-    var res = await checkIntegrity(cid, integrity)
-    if (res){
-      // console.log('Downloaded', manager, name, version, 'from IPFS')
-      await db.put(`cid:${manager}:${name}:${version}`, cid)
-      return true
-      // TODO announce on pubsub (maybe?)
-    } else {
-      console.log('Failed to download', manager, name, version, 'from IPFS', cid)
-      return false
-    }
-  } else {
-    console.log('Unknown version', manager, name, version)
-    return false
-  }
-}
-
-async function goVerify(name, version, cid) {
-  var manager = 'go'
-  // fetch checksum from sumdb
-  var checksum = await go.fetchChecksum(name, version)
-
-  if(!checksum){
-    console.log('Unknown version', manager, name, version)
-    return false
-  }
-
-  // calculate checksum of cid
-  var integrity = await goDigest(version, cid)
-
-  // compare integrity with checksum
-  if(checksum === integrity){
-    await db.put(`cid:${manager}:${name}:${version}`, cid)
-    return true
-  } else {
-    console.log('Failed to download', manager, name, version, 'from IPFS', cid)
-    return false
-  }
-}
-
-async function setTarballSize(manager, name, version) {
-  var cid = await db.get(`cid:${manager}:${name}:${version}`)
-  const { size } = await ipfs.files.stat(`/ipfs/${cid}`)
-  await db.put(`size:${manager}:${name}:${version}`, size)
-  return size
-}
-
-async function getTarballSize(manager, name, version) {
-  await db.get(`size:${manager}:${name}:${version}`)
-}
-
-async function loadMetadata(manager, name) {
-    try {
-      var json = await db.get(`pkg:${manager}:${name}`)
-    } catch (e) {
-      var json = false
-    }
-
-  if(json){
-    // console.log('Loading metadata for', manager, name, 'from cache')
-    return JSON.parse(json)
-  } else {
-    console.log('Loading metadata for', manager, name, 'from registry')
-    return await loadMetadataFromRegistry(manager, name)
-  }
-}
-
-async function loadVersionMetadata(manager, name, version) {
-  var metadata = await loadMetadata(manager, name)
-  if(!metadata) {
-    console.log("Failed to load metadata for", name)
-    return false
-  }
-
-  var versionData = metadata.versions[version]
-
-  if(versionData == null){
-    console.log('Reloading metadata for', name, version)
-    const metadata = await loadMetadataFromRegistry(manager, name)
-    var versionData = metadata.versions[version]
-
-    if(versionData == null){
-      // no known version in registy
-      console.log("Can't find", version, "for", name)
-      return false
-    }
-  }
-  return versionData
-}
-
-// has npm specific code
-async function fetchMetadata(manager, name) {
-  url = "http://registry.npmjs.org/" + name
-  const response = await fetch(url);
-  if (response.ok) {
-    return await response.json();
-  } else {
-    return false
-  }
-}
-
-// has npm specific code
-async function loadMetadataFromRegistry(manager, name) {
-  try{
-    const json = await fetchMetadata(manager, name)
-    if(!json) { return false }
-    await db.put(`pkg:${manager}:${name}`, JSON.stringify(json))
-    return json
-  } catch(e) {
-    console.error("loadMetadataFromRegistry error", manager, name, e)
-    return false
-  }
-}
-
-async function updateMetadata(manager, name) {
-  try{
-    const json = await fetchMetadata(manager, name);
-
-    try {
-      var existingJson = await db.get(`pkg:${manager}:${name}`)
-    } catch(e) {
-      var existingJson = ''
-    }
-
-    if (JSON.stringify(json) !== existingJson) {
-      console.log('Updating', manager, name)
-      await db.put(`pkg:${manager}:${name}`, JSON.stringify(json))
-      return json
-    } else {
-      return false
-    }
-  } catch(e) {
-    return console.error("fetchMetadata error", manager, name, e)
-  }
-}
-
-// has npm specific code
-async function checkIntegrity(cid, integrity) {
-  var sha = ssri.parse(integrity)
-  const responseStream = toStream.readable((async function * () {
-    for await (const chunk of ipfs.cat(cid)) { // cid can't be null here
-      yield chunk.slice()
-    }
-  })())
-
-  var sri = await ssri.fromStream(responseStream, {algorithms: ['sha1', 'sha512']})
-  return !!sri.match(sha)
-}
-
-async function tarballHandler(name, contentType, cid, req, res) {
-  const { size } = await ipfs.files.stat(`/ipfs/${cid}`)
-  const { source } = await detectContentType(name, ipfs.cat(cid)) // TODO don't bother detecting contentType here, it's not reliable
-  const responseStream = toStream.readable((async function * () {
-    for await (const chunk of source) {
-      yield chunk.slice()
-    }
-  })())
-  res.writeHead(200, { 'Content-Type': contentType, 'Content-Length': size });
-  responseStream.pipe(res)
-}
 
 async function validate(manager, name, version) {
   try{
     const cid = await db.get(`cid:${manager}:${name}:${version}`)
 
     if(manager === 'npm'){
-      return await npmVerify(name, version, cid)
+      return await npm.verify(db, name, version, cid)
     } else if (manager === 'go'){
-      return await goVerify(name, version, cid)
+      return await go.verify(db, name, version, cid)
     } else {
       console.log('Unknown manager:', manager)
     }
   } catch(e) {
+    console.error(e)
     return false
   }
 }
 
 async function setupWatcher(callback) {
-  setupNpmWatcher(callback)
-  setupGoWatcher(callback)
-}
-
-function setupNpmWatcher(callback) {
-  const ChangesStream = require("@npmcorp/changes-stream");
-  const changes = new ChangesStream({
-    db: 'https://replicate.npmjs.com/registry',
-    include_docs: true,
-    since: 'now'
-  });
-  changes.on('data', function(change) { callback('npm', change) })
-  changes.on('error', function(error){
-    console.error("Failed to connect to the npm changes feed")
-  })
-}
-
-async function setupGoWatcher(callback) {
-  checkGoPackages(callback)
-  setInterval(async function(){
-    checkGoPackages(callback)
-  }, 60000)
-}
-
-async function checkGoPackages(callback) {
-  try{
-    var currentDate = new Date();
-    var date = new Date(currentDate.getTime() - 60000).toISOString();
-    var res = await fetch(`https://index.golang.org/index?since=${date}`)
-    var text = await res.text()
-    var lines = text.trim().split("\n")
-    lines.forEach((line) => {
-      if(line.length > 0){
-        callback('go', JSON.parse(line))
-      }
-    });
-  } catch(e){
-    console.error("Failed to connect to the go changes feed")
-  }
-}
-
-async function watchNpmImporter(change) {
-  try{
-    var versionNumber = npm.getLatestVersion(change.doc)
-    var version = change.doc.versions[versionNumber]
-    var name = change.doc.name
-
-    try {
-      exists = await db.get(`cid:npm:${name}:${versionNumber}`)
-    } catch (e) {
-      exists = false
-    }
-
-    if(exists) {
-      // duplicate change from stream
-    } else {
-      console.log('New release:', 'npm', name, versionNumber)
-      // if npm and sha512 available, hashAlg = sha2-512
-      addUrltoIPFS('npm', name, versionNumber, version.dist.tarball)
-      loadMetadata('npm', name)
-    }
-  } catch(e){
-    // invalid
-  }
-}
-
-async function watchGoImporter(change) {
-  try{
-    var name = go.escape(change.Path)
-    var version = go.escape(change.Version)
-  } catch(e){
-    console.log('invalid go change', change)
-    console.error(e)
-    return false
-  }
-
-  try {
-    exists = await db.get(`cid:go:${name}:${version}`)
-  } catch (e) {
-    exists = false
-  }
-
-  if(exists) {
-    // duplicate change from stream
-  } else {
-    console.log('New release:', 'go', name, version)
-    fetchGoVersionsList(name)
-    fetchGoLatest(name)
-    loadGoInfo(name, version)
-    loadGoMod(name, version)
-    downloadGoPackageFromRegistry(name, version)
-  }
+  npm.setupWatcher(callback)
+  go.setupWatcher(callback)
 }
 
 async function watchAll() {
   console.log('Watching for new upstream releases for all packages')
   setupWatcher(function(manager, change) {
     if(manager === 'npm'){
-      if(change.doc.name){ watchNpmImporter(change) }
+      if(change.doc.name){ npm.watchImporter(db, change) }
     }
 
     if(manager === 'go'){
-      watchGoImporter(change)
+      go.watchImporter(db, change)
     }
   })
 }
@@ -899,7 +482,7 @@ async function watchKnown() {
         try {
           var known = await db.get(`pkg:${manager}:${name}`)
           if(known){
-            watchNpmImporter(change)
+            npm.watchImporter(change)
           }
         } catch(e){
           // ignore packages we haven't already downloaded
@@ -912,7 +495,7 @@ async function watchKnown() {
       try {
         var known = await db.get(`pkg:${manager}:${name}`)
         if(known){
-          watchGoImporter(change)
+          go.watchImporter(db, change)
         }
       } catch(e){
         // ignore packages we haven't already downloaded
@@ -922,217 +505,40 @@ async function watchKnown() {
 }
 
 async function removePackage(manager, name, version) {
+  // TODO also remove from ipfs
   await db.del(`cid:${manager}:${name}:${version}`)
 }
 
-function filteredReadStream(start) {
-  return db.createReadStream({gte: start, lt: start+ '~'})
-}
-
-async function metadataHandler(name, req, res) {
-  // TODO handle etags and 304 requests
-  // TODO handle npm minimal metadata requests
-  // TODO should probably move saving metadata from res in ProxyRes so it handles private modules
-  const json = await loadMetadata('npm', name)
-
-  res.writeHead(200, {"Content-Type": "application/json"});
-  res.end(JSON.stringify(json));
-}
-
-async function loadGoVersionsList(name) {
-  try {
-    var versions = await db.get(`pkg:go:${name}`)
-  } catch (e) {
-    var versions = false
-  }
-
-  if(versions){
-    return versions
-  } else {
-    return await fetchGoVersionsList(name)
-  }
-}
-
-async function fetchGoVersionsList(name) {
-  var url = `https://proxy.golang.org/${name}/@v/list`
-  const response = await fetch(url);
-  if(response.ok){
-    const body = await response.text();
-    await db.put(`pkg:go:${name}`, body)
-    return body
-  } else {
-    return false
-  }
-}
-
-async function loadGoMod(name, version) {
-  var versionList = await loadGoVersionsList(name)
-  // TODO if version not included in versionList then reload
-
-  try {
-    var versions = await db.get(`mod:go:${name}:${version}`)
-  } catch (e) {
-    var versions = false
-  }
-
-  if(versions){
-    return versions
-  } else {
-    var url = `https://proxy.golang.org/${name}/@v/${version}.mod`
-    var response = await fetch(url);
-    if(response.ok){
-      var body = await response.text();
-      await db.put(`mod:go:${name}:${version}`, body)
-      return body
-    } else {
-      return false
-    }
-  }
-}
-
-async function loadGoInfo(name, version) {
-  var versionList = await loadGoVersionsList(name)
-  // TODO if version not included in versionList then reload
-
-  try {
-    var info = await db.get(`info:go:${name}:${version}`)
-  } catch (e) {
-    var info = false
-  }
-
-  if(info){
-    return JSON.parse(info)
-  } else {
-    var url = `https://proxy.golang.org/${name}/@v/${version}.info`
-    var response = await fetch(url);
-    if(response.ok){
-      var body = await response.json();
-      await db.put(`info:go:${name}:${version}`, JSON.stringify(body))
-      return body
-    } else {
-      return false
-    }
-  }
-}
-
-async function loadGoLatest(name) {
-  var versionList = await loadGoVersionsList(name)
-  // TODO return the latest from versionList
-
-  try {
-    var info = await db.get(`latest:go:${name}`)
-  } catch (e) {
-    var info = false
-  }
-
-  if(info){
-    return JSON.parse(info)
-  } else {
-    return await fetchGoLatest(name)
-  }
-}
-
-async function fetchGoLatest(name) {
-  var url = `https://proxy.golang.org/${name}/@latest`
-  var response = await fetch(url);
-  if(response.ok){
-    var body = await response.json();
-    await db.put(`latest:go:${name}`, JSON.stringify(body))
-    return body
-  } else {
-    return false
-  }
-}
-
-async function getCID(manager, name, version) {
-  await db.get(`cid:${manager}:${name}:${version}`)
-}
-
-async function goDigest(version, cid) {
-  var dir = await tmp.dir()
-
-  var zippath = dir.path + '/' + version + '.zip'
-
-  // get file from ipfs via cid and write to disk
-  await pipe(
-    ipfs.cat(cid),
-    toIterable.sink(fs.createWriteStream(zippath))
-  )
-
-  // digest zip
-  var digest = await zipDigest(zippath)
-
-  // remove tmp directory and zip
-  await fs.remove(zippath)
-
-  return digest
-}
+// general things
 
 async function update(manager, name) {
   if (manager === 'npm'){
-    var update = await updateMetadata('npm', name)
+    var update = await core.updateMetadata('npm', name)
     if (update) {
       var newestVersion = npm.getLatestVersion(update)
-      await downloadPackageFromRegistry('npm', name, newestVersion)
+      await npm.downloadPackageFromRegistry(db, 'npm', name, newestVersion)
     } else {
       return
     }
   } else if (manager === 'go'){
     var versions = await db.get(`pkg:${manager}:${name}`)
-    var newversions = await fetchGoVersionsList(name)
+    var newversions = await go.fetchVersionsList(db, name)
 
     match = versions === newversions
     if (match){
       return
     } else {
-      var latest = await fetchGoLatest(name)
+      var latest = await go.fetchLatest(db, name)
       var version = latest.Version
-      await loadGoInfo(name, version)
-      await loadGoMod(name, version)
-      await downloadGoPackageFromRegistry(name, version)
+      await go.loadInfo(db, name, version)
+      await go.loadMod(db, name, version)
+      await go.downloadPackageFromRegistry(db, name, version)
       return true
     }
   }
 }
 
-async function exportPackages() {
-  await ipfs.files.rm('/forest/export', { recursive: true })
-  await ipfs.files.mkdir('/forest/export', {parents: true})
 
-  var loadpkgs = new Promise((resolve, reject) => {
-    var key = "cid:"
-    var pkgs = []
-    db.createReadStream({gte: key, lt: key+'~'})
-      .on('data', function (data) {
-        var parts = data.key.split(':')
-        var cid = data.value
-        pkgs.push({manager: parts[1], name: parts[2], version: parts[3], cid: cid})
-      })
-      .on('end', function () {
-        resolve(pkgs.sort())
-      })
-  })
-  var packages = await loadpkgs
-
-  for (const pkg of packages) {
-    console.log('Exporting', pkg.manager, pkg.name, pkg.version)
-    try{
-      await ipfs.files.mkdir(`/forest/export/${pkg.manager}/${pkg.name}`, {parents: true})
-    } catch(e) {
-      // folder already exists
-    }
-
-    try{
-      var extension = pkg.manager === 'go' ? 'zip' : 'gzip'
-      await ipfs.files.cp(`/ipfs/${pkg.cid}`, `/forest/export/${pkg.manager}/${pkg.name}/${pkg.version}.${extension}`) // todo gzip or zip
-    } catch(e){
-      // failed to add file
-    }
-  }
-
-  // return root cid of mfs dir
-  return await ipfs.files.stat('/forest/export')
-}
 
 async function search(query) {
   // TODO allow filtering by manager
@@ -1140,49 +546,33 @@ async function search(query) {
   return names.filter(function (pkg) { return pkg.name.indexOf(query) > -1; });
 }
 
+
+
 module.exports = {
   go,
   npm,
-  subscribePackageAnnoucements,
-  unsubscribePackageAnnoucements,
-  addUrltoIPFS,
-  downloadPackageFromIPFS,
-  downloadPackageFromRegistry,
-  listPackages,
-  listPackageNames,
-  checkIntegrity,
-  loadMetadata,
-  reset,
-  loadMetadataFromRegistry,
-  tarballHandler,
-  metadataHandler,
-  validate,
-  connectIPFS,
-  watchKnown,
-  watchAll,
-  updateMetadata,
-  removePackage,
-  closeDB,
-  version,
-  setTarballSize,
-  getTarballSize,
-  filteredReadStream,
-  connectDB,
-  loadGoVersionsList,
-  loadGoInfo,
-  loadGoMod,
-  loadGoLatest,
-  goDigest,
-  getCID,
-  goVerify,
+  core,
+
+  // public
   update,
-  downloadGoPackageFromRegistry,
-  listPeers,
-  exportPackages,
-  dialPeers,
   search,
   seed,
-  loadVersionMetadata,
-  fetchAndAddtoIPFS,
-  guessCID
+  removePackage,
+  subscribePackageAnnoucements,
+  reset,
+  listPackages,
+  listPackageNames,
+  watchAll,
+  watchKnown,
+  validate,
+  importPackage,
+  packageAnnoucementsTopic,
+
+  // not currently internal but should be
+  downloadPackageFromIPFS,
+  connectIPFS,
+  connectDB,
+  closeDB,
+  filteredReadStream,
+  fetchAndAddtoIPFS // used inside republish command
 }
