+++ lib/managers/go.js
@@ -1,6 +1,114 @@
+const tmp = require('tmp-promise');
 const fetch = require('node-fetch')
-const fs = require('fs')
+const fs = require('fs-extra');
 const path = require('path')
+const pipe = require('it-pipe')
+const toIterable = require('stream-to-it')
+const IpfsHttpClient = require('ipfs-http-client')
+const { urlSource } = IpfsHttpClient
+const ipfs = IpfsHttpClient()
+
+const {zipDigest, zipManifest} = require('../zip_digest');
+
+const core = require('./core')
+
+async function serverHandler(db, req, res) {
+  var path = req.url
+
+  if (/@v\/list$/.test(path)){
+    var name = parseName(path)
+    console.log('go list', name)
+
+    var body = await loadVersionsList(db, name)
+
+    if(body){
+      res.writeHead(200, {"Content-Type": "text/plain"});
+      res.end(body);
+    } else {
+      res.writeHead(404);
+      res.end();
+    }
+  }
+
+  if (/@v\/(.+)\.info$/.test(path)){
+    var name = parseName(path)
+    var version = parseVersion(path)
+    console.log('go info', name, version)
+
+    var body = await loadInfo(db, name, version)
+
+    if(body){
+      res.writeHead(200, {"Content-Type": "application/json"});
+      res.end(JSON.stringify(body));
+    } else {
+      res.writeHead(404);
+      res.end();
+    }
+  }
+
+  if (/@v\/(.+)\.mod$/.test(path)){
+    var name = parseName(path)
+    var version = parseVersion(path)
+    console.log('go mod', name, version)
+
+    var body = await loadMod(db, name, version)
+
+    if(body){
+      res.writeHead(200, {"Content-Type": "text/plain"});
+      res.end(body);
+    } else {
+      res.writeHead(404);
+      res.end();
+    }
+  }
+
+  if (/@v\/(.+)\.zip$/.test(path)){
+    var name = parseName(path)
+    var version = parseVersion(path)
+    console.log('go source', name, version)
+
+    var tarball = await returnTarballEarly(db, name, version)
+
+    if(tarball.cid){
+      core.tarballHandler(version+'.zip', 'application/zip', tarball.cid, req, res)
+    } else {
+      downloadPackageFromRegistry(db, name, version)
+      var url = `https://proxy.golang.org/${name}/@v/${version}.zip`
+      // TODO use the zip we just download to ipfs
+      fetch(url).then(resp => new Promise((resolve, reject) => {
+          res.writeHead(resp.status, { 'Content-Type': 'application/zip' });
+          resp.body.pipe(res);
+      }));
+    }
+  }
+
+  if (/@latest$/.test(path)){
+    var name = parseName(path)
+    console.log('go latest', name)
+
+    var body = await loadLatest(name)
+
+    if(body){
+      res.writeHead(200, {"Content-Type": "application/json"});
+      res.end(JSON.stringify(body));
+    } else {
+      res.writeHead(404);
+      res.end();
+    }
+  }
+
+  if(/^\/sumdb/.test(path)){
+    console.log('go sumdb', path)
+    var url = "https://proxy.golang.org"+path
+    const response = await fetch(url);
+    const body = await response.text();
+
+    // TODO handle being offline
+
+    res.writeHead(response.status, {"Content-Type": "text/plain"});
+    res.end(body);
+  }
+}
 
 function parseName(path) {
   return path.match(/\/(.+)\/@/)[1]
@@ -11,7 +119,7 @@ function parseVersion(path) {
   return v.replace('.info', '').replace('.mod', '').replace('.zip', '')
 }
 
-async function returnTarballEarly(name, version) {
+async function returnTarballEarly(db, name, version) {
   try {
     var cid = await db.get(`cid:go:${name}:${version}`)
   } catch (e) {
@@ -60,7 +168,7 @@ function parseGoSum(filepath) {
           resolved: `https://proxy.golang.org/${name}/@v/${version}.zip`,
           integrity: integrity
         }
-        
+
         pkgs.push(pkg)
       }
     }
@@ -83,6 +191,212 @@ function unescape(string) {
   })
 }
 
+async function digest(version, cid) {
+  var dir = await tmp.dir()
+
+  var zippath = dir.path + '/' + version + '.zip'
+
+  // get file from ipfs via cid and write to disk
+  await pipe(
+    ipfs.cat(cid),
+    toIterable.sink(fs.createWriteStream(zippath))
+  )
+
+  // digest zip
+  var digest = await zipDigest(zippath)
+
+  // remove tmp directory and zip
+  await fs.remove(zippath)
+
+  return digest
+}
+
+async function verify(db, name, version, cid) {
+  var manager = 'go'
+  // fetch checksum from sumdb
+  var checksum = await fetchChecksum(name, version)
+
+  if(!checksum){
+    console.log('Unknown version', manager, name, version)
+    return false
+  }
+
+  // calculate checksum of cid
+  var integrity = await digest(version, cid)
+
+  // compare integrity with checksum
+  if(checksum === integrity){
+    await db.put(`cid:${manager}:${name}:${version}`, cid)
+    return true
+  } else {
+    console.log('Failed to download', manager, name, version, 'from IPFS', cid)
+    return false
+  }
+}
+
+async function fetchVersionsList(db, name) {
+  var url = `https://proxy.golang.org/${name}/@v/list`
+  const response = await fetch(url);
+  if(response.ok){
+    const body = await response.text();
+    await db.put(`pkg:go:${name}`, body)
+    return body
+  } else {
+    return false
+  }
+}
+
+async function checkPackages(callback) {
+  try{
+    var currentDate = new Date();
+    var date = new Date(currentDate.getTime() - 60000).toISOString();
+    var res = await fetch(`https://index.golang.org/index?since=${date}`)
+    var text = await res.text()
+    var lines = text.trim().split("\n")
+    lines.forEach((line) => {
+      if(line.length > 0){
+        callback('go', JSON.parse(line))
+      }
+    });
+  } catch(e){
+    console.error("Failed to connect to the go changes feed")
+  }
+}
+
+async function setupWatcher(callback) {
+  checkPackages(callback)
+  setInterval(async function(){
+    checkPackages(callback)
+  }, 60000)
+}
+
+async function loadVersionsList(db, name) {
+  try {
+    var versions = await db.get(`pkg:go:${name}`)
+  } catch (e) {
+    var versions = false
+  }
+
+  if(versions){
+    return versions
+  } else {
+    return await fetchVersionsList(db, name)
+  }
+}
+
+async function loadInfo(db, name, version) {
+  var versionList = await loadVersionsList(db, name)
+  // TODO if version not included in versionList then reload
+
+  try {
+    var info = await db.get(`info:go:${name}:${version}`)
+  } catch (e) {
+    var info = false
+  }
+
+  if(info){
+    return JSON.parse(info)
+  } else {
+    var url = `https://proxy.golang.org/${name}/@v/${version}.info`
+    var response = await fetch(url);
+    if(response.ok){
+      var body = await response.json();
+      await db.put(`info:go:${name}:${version}`, JSON.stringify(body))
+      return body
+    } else {
+      return false
+    }
+  }
+}
+
+async function loadMod(db, name, version) {
+  var versionList = await loadVersionsList(db, name)
+  // TODO if version not included in versionList then reload
+
+  try {
+    var versions = await db.get(`mod:go:${name}:${version}`)
+  } catch (e) {
+    var versions = false
+  }
+
+  if(versions){
+    return versions
+  } else {
+    var url = `https://proxy.golang.org/${name}/@v/${version}.mod`
+    var response = await fetch(url);
+    if(response.ok){
+      var body = await response.text();
+      await db.put(`mod:go:${name}:${version}`, body)
+      return body
+    } else {
+      return false
+    }
+  }
+}
+
+async function loadLatest(name) {
+  var versionList = await loadVersionsList(db, name)
+  // TODO return the latest from versionList
+
+  try {
+    var info = await db.get(`latest:go:${name}`)
+  } catch (e) {
+    var info = false
+  }
+
+  if(info){
+    return JSON.parse(info)
+  } else {
+    return await fetchLatest(db, name)
+  }
+}
+
+async function fetchLatest(db, name) {
+  var url = `https://proxy.golang.org/${name}/@latest`
+  var response = await fetch(url);
+  if(response.ok){
+    var body = await response.json();
+    await db.put(`latest:go:${name}`, JSON.stringify(body))
+    return body
+  } else {
+    return false
+  }
+}
+
+async function downloadPackageFromRegistry(db, name, version) {
+  var url = `https://proxy.golang.org/${name}/@v/${version}.zip`
+  return await core.addUrltoIPFS(db, 'go', name, version, url)
+}
+
+async function watchImporter(db, change) {
+  try{
+    var name = escape(change.Path)
+    var version = escape(change.Version)
+  } catch(e){
+    console.log('invalid go change', change)
+    console.error(e)
+    return false
+  }
+
+  try {
+    exists = await db.get(`cid:go:${name}:${version}`)
+  } catch (e) {
+    exists = false
+  }
+
+  if(exists) {
+    // duplicate change from stream
+  } else {
+    console.log('New release:', 'go', name, version)
+    fetchVersionsList(db, name)
+    fetchLatest(db, name)
+    loadInfo(db, name, version)
+    loadMod(db, name, version)
+    downloadPackageFromRegistry(db, name, version)
+  }
+}
+
+
 module.exports = {
   parseName,
   parseVersion,
@@ -90,5 +404,18 @@ module.exports = {
   fetchChecksum,
   parseGoSum,
   escape,
-  unescape
+  unescape,
+  digest,
+  verify,
+  fetchVersionsList,
+  checkPackages,
+  setupWatcher,
+  loadVersionsList,
+  loadInfo,
+  loadMod,
+  loadLatest,
+  fetchLatest,
+  downloadPackageFromRegistry,
+  watchImporter,
+  serverHandler
 }
