+++ lib/managers/npm.js
@@ -1,5 +1,12 @@
 const validateNpmPackageName = require("validate-npm-package-name")
 
+const core = require('./core')
+const ssri = require('ssri')
+const toStream = require('it-to-stream')
+const IpfsHttpClient = require('ipfs-http-client')
+const { urlSource } = IpfsHttpClient
+const ipfs = IpfsHttpClient()
+
 function isTarballRequest(path) {
   if (path.match('\.tgz$')) {
     parts = path.split("/")
@@ -60,11 +67,140 @@ function getLatestVersion(doc) {
   return latest[0]
 }
 
+async function verify(db, name, version, cid) {
+  var manager = 'npm'
+  var metadata = await core.loadMetadata(db, manager, name)
+  if(!metadata) {
+    console.log("Failed to load metadata for", name)
+    return false
+  }
+  var versionData = metadata.versions[version]
+
+  if (versionData == null) {
+    console.log('Reloading metadata for', name, version)
+    var metadata = await core.loadMetadataFromRegistry(db, manager, name)
+    var versionData = metadata.versions[version]
+  }
+
+  if (versionData) {
+    if(versionData.dist.integrity){
+      var integrity = versionData.dist.integrity
+    } else {
+      var integrity = ssri.fromHex(versionData.dist.shasum, 'sha1').toString()
+    }
+    var res = await checkIntegrity(cid, integrity)
+    if (res){
+      // console.log('Downloaded', manager, name, version, 'from IPFS')
+      await core.saveCid(db, manager, name, version, cid)
+      return true
+      // TODO announce on pubsub (maybe?)
+    } else {
+      console.log('Failed to download', manager, name, version, 'from IPFS', cid)
+      return false
+    }
+  } else {
+    console.log('Unknown version', manager, name, version)
+    return false
+  }
+}
+
+async function checkIntegrity(cid, integrity) {
+  var sha = ssri.parse(integrity)
+  const responseStream = toStream.readable((async function * () {
+    for await (const chunk of ipfs.cat(cid)) { // cid can't be null here
+      yield chunk.slice()
+    }
+  })())
+
+  var sri = await ssri.fromStream(responseStream, {algorithms: ['sha1', 'sha512']})
+  return !!sri.match(sha)
+}
+
+function setupWatcher(callback) {
+  const ChangesStream = require("@npmcorp/changes-stream");
+  const changes = new ChangesStream({
+    db: 'https://replicate.npmjs.com/registry',
+    include_docs: true,
+    since: 'now'
+  });
+  changes.on('data', function(change) { callback('npm', change) })
+  changes.on('error', function(error){
+    console.error("Failed to connect to the npm changes feed")
+  })
+}
+
+async function watchImporter(db, change) {
+  try{
+    var versionNumber = getLatestVersion(change.doc)
+    var version = change.doc.versions[versionNumber]
+    var name = change.doc.name
+
+    try {
+      exists = await db.get(`cid:npm:${name}:${versionNumber}`)
+    } catch (e) {
+      exists = false
+    }
+
+    if(exists) {
+      // duplicate change from stream
+    } else {
+      console.log('New release:', 'npm', name, versionNumber)
+      // if npm and sha512 available, hashAlg = sha2-512
+      core.addUrltoIPFS(db, 'npm', name, versionNumber, version.dist.tarball)
+      core.loadMetadata(db, 'npm', name)
+    }
+  } catch(e){
+    // invalid
+    console.error(e)
+  }
+}
+
+async function downloadPackageFromRegistry(db, manager, name, version){
+  var metadata = await core.loadMetadata(db, manager, name)
+  if(!metadata) {
+    console.log("Failed to load metadata for", name)
+    return false
+  }
+
+  var versionData = metadata.versions[version]
+
+  if(versionData == null){
+    console.log('Reloading metadata for', name, version)
+    const metadata = await core.loadMetadataFromRegistry(db, manager, name)
+    var versionData = metadata.versions[version]
+
+    if(versionData == null){
+      // no known version in registy
+      console.log("Can't find", version, "for", name)
+      return false
+    }
+  }
+
+  var url = versionData.dist.tarball
+  // if sha512 available, hashAlg = sha2-512
+  return await core.addUrltoIPFS(db, manager, name, version, url)
+}
+
+async function metadataHandler(name, req, res) {
+  // TODO handle etags and 304 requests
+  // TODO handle npm minimal metadata requests
+  // TODO should probably move saving metadata from res in ProxyRes so it handles private modules
+  const json = await core.loadMetadata(db, 'npm', name)
+
+  res.writeHead(200, {"Content-Type": "application/json"});
+  res.end(JSON.stringify(json));
+}
+
 module.exports = {
   getLatestVersion,
   isTarballRequest,
   isMetadataRequest,
   returnTarballEarly,
   setConfig,
-  removeConfig
+  removeConfig,
+  verify,
+  setupWatcher,
+  downloadPackageFromRegistry,
+  metadataHandler,
+  watchImporter
 }
